server admin: Avoid having crawlers mutate server state.  (#2297)
*Description*:
Keeps track of whether admin handlers mutate state or not.  When
showing the root page, use a form with a button that POSTs, rather
than an a-tag referencing a URL that mutates state.

Also: in preparing the sorted-list of handlers, use a std::vector and
std::sort rather than using a map to get the sorting behavior, as it's
generally faster, since (a) it doesn't have to keep the collection
sorted as its being incrementally inserted and (b) it doesn't have to
create as many internal structures with pointers.  This doesn't matter
for this handler but I thought I'd taste-test the pattern.

*Risk Level*: Low

*Testing*:
manually brought up server and tried admin page.
bazel test //test/...
added new unit test to make sure mutators get forms and non-mutators get a-tags.

*Docs Changes*:
none

*Release Notes*:
none

Fixes https://github.com/envoyproxy/envoy/issues/2296

Signed-off-by: Joshua Marantz <jmarantz@google.com>