load balancer: implement Maglev algorithm and benchmarks (#2574)
This commit implements the Maglev consistent hash algorithm and a benchmark
around it to compare it to RingHash. This is not a full load balancer implementation
but it is enough to get an idea of the rough relative performance. The one
missing piece of data is a test/simulation of how many keys move when a host
is removed. The number of moved keys will be higher with Maglev than with Ketama.

Signed-off-by: Matt Klein <mklein@lyft.com>