server: un-revert #2568 with a fix to initialize logging before validation (#2623)
Description:
Reverts #2613 the revert of #2568
Fixed the underlying issue which was that logging was not initialized during validation, so it ran lockless. This caused intermittent tsan errors. Although #2619 makes the failure immediate and consistent by asserting that logging has been initialized prior to spawning any threads.

Risk Level: Medium -- the earlier #2568 caused intermittent tsan errors which we believe to be fixed, but #2568 was medium-risk in the first place.

Release Notes: N/A