Support for external authorization tcp filter. (#2416)
As per the feedback on #2359 I am breaking it up into three PR's.

The patch in this PR adds support for the TCP filter.
This PR implements the discussing we had in #2291.

Testing
I have tested this on my local setup with an external gRPC authorization server. I have also added UT's for the TCP filter.

Risk
Low: Because only the users of this filter will be impacted. It should not impact general stability of Envoy it self.

Signed-off-by: Saurabh Mohan <saurabh+github@tigera.io>