Upgrade to http-parser 2.8.0. (#2720)
http-parser doesn't provide a changelog. The commits between releases are here. The biggest change is how http-parser parses HTTP version numbers. The code now adheres to the RFC which says the major and minor version is a single digit. Previously Envoy would test a low HTTP version number of 0.8 which would fail as expected as http-parser would reject major version numbers below 1 (even though technically 0.9 is a valid HTTP version). It was confirmed on Twitter (thread) with @indutny that with the new change 0.8 is considered valid. In addition to this dependency change we change the test to test for an invalid version.

Risk Level: Medium. Changes to HTTP version parsing. Shouldn't affect Envoy with regular 1.1 requests however 0.8 is now considered valid as per HTTP RFC.

Testing: bazel test //test/... and running on local instances for 1+ weeks.

Docs Changes: None required.

Signed-off-by: Michael Payne <michael@sooper.org>