Support for external authorization http filter. (#2417)
Authorization TCP and HTTP filters using an external gRPC service.
Patch 2 of 3: HTTP filter.

Description
As per the feedback on #2359 I am breaking it up into three PR's.

The patch in this PR adds support for the HTTP filter.
This PR implements the discussing we had in #2291.

Testing
I have tested this on my local setup with an external gRPC authorization server. I have also added UT's for the HTTP filter.

Risk
Low: Because only the users of this filter will be impacted. It should not impact general stability of Envoy it self.

Signed-off-by: Saurabh Mohan <saurabh+github@tigera.io>