upstream: log metadata used to create subset lbs (#2772)
Per a slack conversation, it would be useful is the subset load balancer logged the metadata it used to create it's underlying loadbalancers. This modifies the subset load balancer to track the default subset using the same data type used for non-default subsets (thereby eliminating the `hostMatchesDefaultSubset` function). Logging occurs at the debug level.

*Risk Level*: Low

*Testing*: existing unit tests cover the aded code and prove that no functional changes occurred.

Signed-off-by: Stephan Zuercher <stephan@turbinelabs.io>