util: Adds some basic unit tests for check_format.py. (#3495)
Description:

Adds a bit of unit-testing to the format fixer/checker script, which has grown complex. These unit tests are by no means comprehensive but I thought I'd get feedback on the pattern before going much further. I actually don't have a huge amount of experience writing gunit tests in python, and it looks like there's currently no gunit tests in Python in the Envoy codebase. But I'm willing to switch this to gunit if we think that's useful.

The underlying python test program MUST be run under docker, so it's got an awkward name and comes with a bash script to run it under docker directly.

IMO check_format.py has grown organically and needs to be refactored. I have such a refactor pending, but it will be a lot easier to get confidence in the refactor when there are tests.

Once the script is tested and refactored, it will be trivial to add detection of std::mutex and std::condition_variable, which is the ultimate goal.

Risk Level: Low

Testing: tools/check_format.py -- to make sure this doesn't introduce format errors itself.
Manually ran test/check_format_test.sh.

Docs Changes: N/A
Release Notes: N/A

Signed-off-by: Joshua Marantz <jmarantz@google.com>