Revert "stats: envoy prometheus endpoint fails promlint due to format issue  #3019 (#3049)" (#3830)
This reverts commit 64dcf13b05af19d93a9581bdfc46234ddd8050d8.

The originally reported bug (#3019) was due to metric names not being sanitized, which I think was resolved by #2303.
Change #3049, which was intended to fix #3019, erroneously applied sanitization to label values instead. Prometheus label values are permitted to contain any Unicode char, so this change was unnecessary and leads to an unexpected change in e.g. cluster names.

This PR reverts the change, so that - and . will be preserved in label values.

Risk Level: low

Signed-off-by: Jonathan Oddy <jonathan.oddy@transferwise.com>