fuzz: reduce log level when running under fuzz engine. (#4161)
The ClusterFuzz stats indicated a lot of log spew, this is a structural
attempt to reduce it. We now run with logging off when under the engine,
but at ::info when running under bazel test. This can be further
overrides when debugging with -l to more verbose levels. The usual flow
is we download a repeater from the database and then reproduce with
logs.

There's still some log spam from libproto, will look into that
separately.

Risk Level: low
Testing: manual tests on CLI with different combinations of flags.

Signed-off-by: Harvey Tuch <htuch@google.com>