Creating JwksFetcher interface and impl v2 (#4242)
JwksFetcher wraps up HTTP acquisition of JWKS strings converting them into a concrete type on the way.
JwksFetcher is reusable so can be used in a wider context.

Tests updated and fixed where necessary.

We are in the process of implementing a new Envoy filter based on the design presented here and wish to re-use existing logic in the jwt_authn filter. We've split out the logic we're interested in into a new class called JwksFetcher. Later PRs will re-use the split out logic an OpenID Connect filter.

This is the second crack at this PR (see #4225 which went horribly wrong after following the DCO rebase guidelines).

Risk Level: Medium
Testing: Add replacement and additional unit tests for the logic that's been moved.

Signed-off-by: Nick A. Smith <nick.a.smith@thales-esecurity.com>