Revert 41304e7f510089e6745647078baf9e9a95f50f70 (#4581)
This PR reverts #4382. When deploying at Lyft we noticed crashes on here where we might be derefencing the connection_stats_ pointer after the point has been reset.

Note: this PR keeps the changes to the API made in the original PR but tags the field as not implemented. This is what we have done in the past for reverts that involve changes that change the API.

Signed-off-by: Jose Nino <jnino@lyft.com>