cds: continue with remaining clusters if one addOrUpdateCluster fails (#5806)
Currently when multiple clusters are changed with CDS, the first cluster in error aborts the application of these changes. This means that all changes before a cluster with an error are applied, but all changes after that cluster are not. This fix changes the behaviour so that exceptions in addOrUpdateCluster are collected and reported after the loop over all clusters.

Risk Level: Low
Testing: We did a manual test with the change in place, unit tests added.

Co-authored-by: Jakob Schmid <jakob.schmid@sap.com>
Signed-off-by: Axel Siebenborn <axel.siebenborn@sap.com>