Wrap std::string constructor around pair argument. (#6070)
Description: This is a bad PR, and I feel bad. Perhaps related to the finally occurring deprecation of our string troubles at Google, we have a compilation failure around these lines because `absl::string_view` to `std::string` conversion is broken in this subtle way when used to initialize an `std::pair` with a string member. Explicitly call the `std::string` constructor to avoid this compilation error. Hopefully we can remove this hack in about a month when we can finally clean up the rest of the silly string hacks.

Risk Level: Low
Testing: bazel test //test/...
Docs Changes: None
Release Notes: None

Signed-off-by: Dan Noé <dpn@google.com>