Allow any host when fallback policy fails (#6117)
We use the default subset as our fallback policy. However, if there's
a problem with our metadata it's possible for the default subset
to become empty. In this case, we'd like any host to be selected.

This change adds a panic_mode_any boolean option to LbSubsetConfig
to enable this behavior.

Signed-off-by: Raul Gutierrez Segales <rgs@pinterest.com>