Add support to unregister from lifecycle notifications (#6984)
Return a handle from registration methods that can be used to unregister. This is needed for registrations made by objects that can have a shorter lifetime than the server (such as filters or access logs). When those objects are destroyed the callbacks will be invalidated so the registrations should be removed to avoid memory access violations.

Risk Level: low
Testing: unit tests

Signed-off-by: Elisha Ziskind <eziskind@google.com>