sds: add validation callback for all secret providers (#8310)
This adds an implementation of addValidationCallback for all secret providers. This is a no-op, unless the secret provider is a `CertificateValidationContextSdsApi.

Envoy crashes when trying to add a validation callback to CertificateValidationContextProvider due to a bad dynamic cast. When certificate_validation_context_provider_ isn't a CertificateValidationContextSdsApi, we no-op.

Fixes: #7892 and OSS-Fuzz Issues
https://bugs.chromium.org/p/oss-fuzz/issues/detail?id=17001
https://bugs.chromium.org/p/oss-fuzz/issues/detail?id=16766
Testing: add corpus entry

Signed-off-by: Asra Ali <asraa@google.com>