upstream: fix eds crash when changing priorities (#9368)
When receiving an EDS config update, if there is an empty priority in the new update, the update function gets skipped instead of updating the priority with an empty list of hosts. This is particularly bad when the update would move an existing host to a different priority, because the result is the host ends up in two priorities because it doesn't get removed from the existing priority. This change ensures that the update function is called with an empty list of hosts if the priority was previously configured with hosts and the new config does not contain hosts in that priority.

Signed-off-by: michael.wozniak <michael.wozniak@snapchat.com>