server: add getTransportSocketFactoryContext() to Filter context (#9211)
Add getTransportSocketFactoryContext() that returns a TransportSocketFactoryContext with a lifetime at least as long as the one of the server. This can be used by extensions to create ssl
contexts that can be shared across listeners.

Implementation of this required harmonization of method names and constness of some of the methods in class TransportSocketFactoryContext, so that implementing classes would not
need to re-implement similar methods with different names.

ServerFactoryContextImpl is extended to implement TransportSocketFactoryContext in addition to ServerFactoryContext; this allows for minimal implementation and avoids duplication of code. It may might be possible to combine these interfaces as a future enhancement.

The existing TransportSocketFactoryContextImpl is still used as before, as it allows for listener specific stats scopes and listener specific init manager to be used. TransportSocketFactoryContext returned by the new getTransportSocketFactoryContext() returns Server's stats scope and the Server's init manager.

Risk Level: Low (existing use of TransportSocketFactoryContextImpl is retained)
Testing: external repo (github.com/cilium/proxy) uses this.
Docs Changes: none
Release Notes: none

Signed-off-by: Jarno Rajahalme <jarno@covalent.io>