upstream: load distribution in Total Panic mode (#9343)
Changed how load is calculated when all priority levels are in panic mode. Each priority level receives percentage of the traffic related to the number of hosts in that priority regardless of the health status of the hosts. This smooths out how traffic is shifted when hosts become unhealthy. See #4685 for design proposal and discussion.

Signed-off-by: Christoph Pakulski <paker8848@gmail.com>