adaptive concurrency: Fix min_concurrency behavior when gradient shrinks (#9947)
The min_concurrency setting dicated the concurrency limit when in a
min_rtt calculation window. However, if the gradient caused the
concurrency limit to shrink, it was possible for the limit to drop below
this configured minimum. This patch fixes this behavior so that the
calculated concurrency limit is always >= the configured minimum.

In addition, a minor change to the stat indicating whether the min_rtt
calculation window is active. When active, the stat was being set to the
concurrency limit; however, according to the documentation this should
be either 0 or 1. This patch also fixes this behavior.

Signed-off-by: Tony Allen <tony@allen.gg>