runtime: Add support for setting default value when parsing string value (#10058)
This change adds the ability to set a default value when calling get(...) in runtime. Prior to this change, if there was no entry matching the key, an empty string was returned. This change is to make the return value more flexible.

Risk Level: Low
Testing: Changed existing unit test result
Docs Changes: N/A
Release Notes: N/A

Signed-off-by: Lisa Lu <lisalu@lyft.com>