CacheFilter: add parsing of range requests (#11943)
Tackles the first part of issue #10132 (parsing range requests). Added a function getRequestRanges to parse ranges from requests' headers into LookupRequest. This allows the creation of appropriate ranges for the LookupResult object, which will have the corresponding bytes for the content of the response. In turn, this will allow the extension of the function CacheFilter::onHeaders to provide only the requested range(s) of content.

Signed-off-by: Caio <caiomelo@google.com>
Co-authored-by: Josiah Kiehl <kiehl@google.com>