Windows: Fix mysql tests (#12732)
Using a union to encode mysql header relied on undefined behavior of
unions in C++: "It's undefined behavior to read from the member of the
union that wasn't most recently written. Many compilers implement, as a
non-standard language extension, the ability to read inactive members of
a union." In conjunction with the `bits_` field of the union not being
initialized, this caused invalid sequence numbers to be parsed. Replaced
usage of union with bit shift/mask.

Signed-off-by: Sunjay Bhatia <sunjayb@vmware.com>
Co-authored-by: William A Rowe Jr <wrowe@vmware.com>