stats: Add API Scope::scopeFromStatName and use it to skip some symbol table locks during routing recompute from control-plane update. (#14418)
Commit Message: Adds an API for constructing scopes from StatName without holding a string representation, which helps avoid taking a lock. It was previously thought that scopes were not created on the hot-path, so this was presumed not to be necessary. But flame-graphs and code analysis suggest we can still contend on VirtualCluster creation, which involves making scopes.

As of now we are not trying to remove all creation of scopes from strings, but that is something we may want to consider as a follow-up.
Additional Description:
Risk Level: medium
Testing: //test/...
Docs Changes: n/a
Release Notes: n/a
Platform Specific Features: n/a

Signed-off-by: Joshua Marantz <jmarantz@google.com>